requirements.txt
# Core dependencies
tiktoken>=0.5.0
nltk>=3.8.0
numpy>=1.24.0

# Embeddings and transformers
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# PDF processing
pymupdf>=1.23.0
langchain>=0.0.300
langchain-text-splitters>=0.0.1

# Database
pymilvus>=2.3.0

# Web interface
gradio>=3.40.0

# Utilities
python-dotenv>=1.0.0

# LLM APIs
groq>=0.4.0
